requirements.txt
requests>=2.28.0
polars>=0.20.0
pyarrow>=14.0.0
python-dateutil>=2.8.0
